requirements.txt
google-generativeai==0.2.1
streamlit==1.28.1
streamlit-lottie==0.0.5
google-auth==2.23.4
nltk==3.8.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pix2tex==0.1.2
transformers==4.35.0
torch==2.1.0
keras==2.15.0
xgboost==1.7.3
joblib==1.2.0
numpy==1.24.3
pandas==2.0.3
protobuf==4.25.0
python-dotenv==1.0.0
Request==2.31.0
scikit-learn==1.3.0
scipy==1.11.3